Misc. scheduler performance improvements
Summary:
Added benchmark for concurrent torrent downloads from multiple peers, and profiled
cpu usage / mutex contention in scheduler, making the following changes:

- Removed unnecessary FileEntry locking on Close.
- Removed `log.Debug` statements from hot code paths (this actually acquires a lock, even if debug is disabled).
- Add on-disk piece completion tracking to make torrent initialization cheaper (it happens on every incoming handshake).

Before:

```
goos: darwin
goarch: amd64
pkg: code.uber.internal/infra/kraken/lib/torrent/scheduler
BenchmarkPieceUploadingAndDownloading-8                1        51544198352 ns/op
PASS
ok      code.uber.internal/infra/kraken/lib/torrent/scheduler   70.735s
```

After:

```
goos: darwin
goarch: amd64
pkg: code.uber.internal/infra/kraken/lib/torrent/scheduler
BenchmarkPieceUploadingAndDownloading-8                1        35272844693 ns/op
PASS
ok      code.uber.internal/infra/kraken/lib/torrent/scheduler   53.670s
```

Reviewers: #kraken, yiran

Reviewed By: #kraken, yiran

Subscribers: yiran, jenkins

Differential Revision: https://code.uberinternal.com/D1326475

+++ lib/torrent/scheduler/conn_test.go
@@ -16,13 +16,13 @@ func TestSetEgressBandwidthThrottlesPieceSending(t *testing.T) {
 
 	size := 4 * memsize.KB
 	numPieces := 256
-	pieceLength := int(size / uint64(numPieces))
+	pieceLength := size / uint64(numPieces)
 	bytesPerSec := memsize.KB
 	expectedDur := time.Duration(size/bytesPerSec) * time.Second
 
 	config := connConfigFixture()
 
-	c, cleanup := connFixture(t, config, pieceLength)
+	c, cleanup := connFixture(t, config, int(pieceLength))
 	defer cleanup()
 
 	c.SetEgressBandwidthLimit(bytesPerSec)
